Move FAQ component to make it re-usable by other pages.
